feat(queue): retries, dead letter queue, idempotency, leases, concurrency
The worker previously had none of the guarantees it appeared to have: a job
claimed by a worker that then died stayed in 'processing' forever, failures
were terminal with no retry, and jobs ran one at a time.

- retries with exponential backoff up to max_attempts, then status 'dead'
- reaper returns expired leases to the queue so a job is never stranded
- idempotency key deduplicates publishes via a partial unique index
- delayed jobs via available_at
- bounded in-process concurrency with a semaphore
- poll interval as a fallback for a missed NOTIFY
- queue-admin bin: `just queue-status`, `just queue-redrive <channel>`
- graceful shutdown for api and worker on SIGTERM/Ctrl+C, closing pools
- SeaORM connection pool options instead of defaults
- per-app .env and .env.example; no root .env
- 5 behaviour tests covering retry, DLQ, lease expiry, skip-locked, delay

Delivery stays at-least-once by design, so consumers must be idempotent.

+++ apps/api/src/modules/example/dto.rs
@@ -58,7 +58,7 @@ pub struct BulkResultResponse {
 
 #[derive(Debug, Serialize, ToSchema)]
 pub struct PublishJobResponse {
-    pub job_id: i32,
+    pub job_id: i64,
 }
 
 #[derive(Debug, Deserialize, IntoParams)]
